backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
chromadb
sentence-transformers
requests
PyPDF2
python-dotenv